more int2str type correctness fixing

+++ apps/dsm/mods/mod_mysql/ModMysql.cpp
@@ -254,9 +254,9 @@ EXEC_ACTION_START(SCMyExecuteAction) {
     mysqlpp::SimpleResult res = query.execute();
     if (res) {
       sc_sess->CLR_ERRNO;
-      sc_sess->var["db.rows"] = int2str(res.rows());
+      sc_sess->var["db.rows"] = int2str((long)res.rows());
       sc_sess->var["db.info"] = res.info();
-      sc_sess->var["db.insert_id"] = int2str(res.insert_id());
+      sc_sess->var["db.insert_id"] = int2str((long)res.insert_id());
     } else {
       sc_sess->SET_ERRNO(DSM_ERRNO_MY_QUERY);
       sc_sess->SET_STRERROR(res.info());
@@ -295,7 +295,7 @@ EXEC_ACTION_START(SCMyQueryAction) {
       sc_sess->transferOwnership(m_res);
 
       sc_sess->CLR_ERRNO;    
-      sc_sess->var["db.rows"] = int2str(res.num_rows());
+      sc_sess->var["db.rows"] = int2str((unsigned int)res.num_rows());
     } else {
       sc_sess->SET_ERRNO(DSM_ERRNO_MY_QUERY);
       sc_sess->SET_STRERROR("query did not have a result");
@@ -344,7 +344,7 @@ EXEC_ACTION_START(SCMyQueryGetResultAction) {
       }
 
       sc_sess->CLR_ERRNO;    
-      sc_sess->var["db.rows"] = int2str(res.num_rows());
+      sc_sess->var["db.rows"] = int2str((unsigned int)res.num_rows());
     } else {
       sc_sess->SET_ERRNO(DSM_ERRNO_MY_QUERY);
     }
@@ -591,9 +591,9 @@ EXEC_ACTION_START(SCMyPutFileToDBAction) {
     mysqlpp::SimpleResult res = query.execute();
     if (res) {
       sc_sess->CLR_ERRNO;
-      sc_sess->var["db.rows"] = int2str(res.rows());
+      sc_sess->var["db.rows"] = int2str((long)res.rows());
       sc_sess->var["db.info"] = res.info();
-      sc_sess->var["db.insert_id"] = int2str(res.insert_id());
+      sc_sess->var["db.insert_id"] = int2str((long)res.insert_id());
     } else {
       sc_sess->SET_ERRNO(DSM_ERRNO_MY_QUERY);
       sc_sess->SET_STRERROR(res.info());
